Move iterateTree to helpers and add a test
